fix sale_date and create.blade

+++ database/migrations/2023_06_29_134051_create_comics_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('title', 100);
             $table->text('description');
             $table->string('thumb', 350)->nullable();
-            $table->tinyInteger('price')->unsigned();
+            $table->decimal('price', 4, 2)->unsigned();
             $table->string('series', 100);
             $table->date('sale_date');
             $table->string('type', 100);
